+++ packages/jacaranda/src/middlewares/index.js
@@ -1,3 +1,4 @@
 export { default as jsonError } from './jsonError';
 export { default as favicon } from './favicon';
 export { default as action } from './action';
+export { default as requestId } from './requestId';